refactor: move cascade resolver from gene-app to atlas-browser plugin
Move ModelAtlasClient usage and atlasURIConverter into atlas-browser plugin.
Register as TSM service 'gene.atlas.cascadeResolver' with configure() method.
App.vue calls the service instead of importing storage-model-atlas directly.

gene-app no longer depends on storage-model-atlas.
