+++ packages/atlas-browser/src/composables/atlasURIConverter.ts
